Fix install error on non-existent .gimp-2.8

+++ tools/install.sh
@@ -47,9 +47,11 @@ main() {
 
   gimp_ps_directory="$HOME/.gimp-$version"
 
-  # Backup previous directory
-  now=$(date +"%Y%m%d%H%M%S")
-  mv "$gimp_ps_directory" "$gimp_ps_directory.backup.$now"
+  # Backup previous directory, if any
+  if [ -e "$gimp_ps_directory" ]; then
+    now=$(date +"%Y%m%d%H%M%S")
+    mv "$gimp_ps_directory" "$gimp_ps_directory.backup.$now"
+  fi
 
   env git clone --depth=1 https://github.com/doctormo/GimpPs.git $gimp_ps_directory || {
     printf "Error: git clone of oh-my-zsh repo failed\n"
